requirements.txt
asgiref==3.8.1
Django==5.1.4
django-filter==24.3
django-stubs==5.1.1
django-stubs-ext==5.1.1
djangorestframework==3.15.2
isort==5.13.2
sqlparse==0.5.3
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
